Refactor database package structure and rename GameLinkToDatabase to GameLinkToDatabaseHelper
- Consolidate database interaction classes into a single `com.example.playcheck.database` package.
- Rename `GameLinkToDatabase` to `GameLinkToDatabaseHelper` and update all references in `GameList.java`.
- Update imports across multiple activities and models to reflect the new package structure.
- Clean up unused imports and update AGP version in `libs.versions.toml`.
