--- app/src/main/java/com/example/playcheck/Database/GameLinkToDatabase.java
@@ -1,5 +1,0 @@
-package com.example.playcheck.Database;
-
-public class GameLinkToDatabase {
-
-}
