--- app/src/main/java/com/example/playcheck/Database/TeamLinkToDatabase.java
@@ -1,4 +1,0 @@
-package com.example.playcheck.Database;
-
-public class TeamLinkToDatabase {
-}
