+++ app/src/main/java/com/example/playcheck/database/GameLinkToDatabase.java
@@ -1,0 +1,5 @@
+package com.example.playcheck.database;
+
+public class GameLinkToDatabase {
+
+}
