+++ app/src/main/java/com/example/playcheck/database/TeamLinkToDatabase.java
@@ -1,0 +1,4 @@
+package com.example.playcheck.database;
+
+public class TeamLinkToDatabase {
+}
