+++ app/src/main/java/com/example/playcheck/puremodel/Referee.java
@@ -1,6 +1,6 @@
 package com.example.playcheck.puremodel;
 
-import com.example.playcheck.Database.RefereeLinkToDatabase;
+import com.example.playcheck.database.RefereeLinkToDatabase;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
