+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "9.0.0"
+agp = "9.1.0"
 firebaseBom = "34.9.0"
 junit = "4.13.2"
 junitVersion = "1.3.0"
